refactor(gallery/admin): drop "admin-" prefix from demo imports (#930, bug-3)
Revert the workaround for the manifest-key collision: now that the build
script keys nested components by path, the gallery demos don't need a
unique "admin-" prefix on their filenames. Update the pages to import
from the unprefixed paths.
